Reorganized component folders, and moved header into switch component
