+++ src/features/home-page/HomePage.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import HeroComponent from '../../components/home-page/Hero';
 
 const HomePage = () => {
     return (
-        <main className='w-all'>
-            
+        <main className='w-all flex column justify-start'>
+            <HeroComponent />
         </main>
     )
 }
